requirements.txt
langchain
langchain-google-genai
langchain-chroma
langchain-community
langchain-text-splitters
chromadb
pypdf
python-dotenv
streamlit

